+++ src/main/java/org/openlmis/requisition/service/JasperReportsViewService.java
@@ -39,7 +39,7 @@ import org.openlmis.requisition.domain.Requisition;
 import org.openlmis.requisition.domain.RequisitionStatus;
 import org.openlmis.requisition.domain.RequisitionTemplate;
 import org.openlmis.requisition.domain.RequisitionTemplateColumn;
-import org.openlmis.requisition.dto.BasicFacilityDto;
+import org.openlmis.requisition.dto.MinimalFacilityDto;
 import org.openlmis.requisition.dto.FacilityDto;
 import org.openlmis.requisition.dto.GeographicZoneDto;
 import org.openlmis.requisition.dto.ProcessingPeriodDto;
@@ -342,16 +342,16 @@ public class JasperReportsViewService {
         .filter(RequisitionStatus::isApproved)
         .collect(Collectors.toSet());
 
-    List<BasicFacilityDto> facilities;
+    List<MinimalFacilityDto> facilities = new ArrayList<>();
     if (district != null) {
-      facilities = facilityReferenceDataService.search(null, null, district.getId(), true);
+      facilities.addAll(facilityReferenceDataService.search(null, null, district.getId(), true));
     } else {
-      facilities = facilityReferenceDataService.basicFindAll();
+      facilities.addAll(facilityReferenceDataService.findAll());
     }
 
     List<TimelinessReportFacilityDto> facilitiesMissingRnR = new ArrayList<>();
     // find active facilities that are missing R&R
-    for (BasicFacilityDto facility : facilities) {
+    for (MinimalFacilityDto facility : facilities) {
       if (facility.getActive()) {
         List<Requisition> requisitions = requisitionService.searchRequisitions(
             facility.getId(), program.getId(), null, null, processingPeriod.getId(),
@@ -365,9 +365,9 @@ public class JasperReportsViewService {
     }
 
     // sort alphabetically by district and then facility name
-    Comparator<BasicFacilityDto> comparator = Comparator.comparing(
+    Comparator<MinimalFacilityDto> comparator = Comparator.comparing(
         facility -> facility.getZoneByLevelNumber(DISTRICT_LEVEL).getName());
-    comparator = comparator.thenComparing(Comparator.comparing(BasicFacilityDto::getName));
+    comparator = comparator.thenComparing(Comparator.comparing(MinimalFacilityDto::getName));
 
     return facilitiesMissingRnR.stream()
         .sorted(comparator).collect(Collectors.toList());